Split routes + minor changes
