+++ views/addSingle.ejs
@@ -50,21 +50,21 @@
       <div class="form-row">
         <div class="form-group">
           <label for="sellkeys">Sell keys</label>
-          <input type="number" class="form-control" id="sellkeys" name="sellkeys" min="0" value="0">
+          <input type="number" class="form-control" id="sellkeys" name="sellkeys" min="0" value="0" required>
         </div>
         <div class="form-group">
           <label for="sellmetal">Sell metal</label>
-          <input type="number" class="form-control" id="sellmetal" name="sellmetal" min="0" step="any">
+          <input type="number" class="form-control" id="sellmetal" name="sellmetal" min="0" step="any" required>
         </div>
       </div>
       <div class="form-row">
         <div class="form-group">
           <label for="buykeys">Buy keys</label>
-          <input type="number" class="form-control" id="buykeys" name="buykeys" min="0"value="0">
+          <input type="number" class="form-control" id="buykeys" name="buykeys" min="0"value="0" required>
         </div>
         <div class="form-group">
           <label for="buymetal">Buy metal</label>
-          <input type="number" class="form-control" id="buymetal" name="buymetal" min="0" step="any">
+          <input type="number" class="form-control" id="buymetal" name="buymetal" min="0" step="any" required>
         </div>
       </div>
     </div>
