+++ views/index.ejs
@@ -165,7 +165,7 @@
     <% if (pricelist.length !== 0) { %>
     <br>
 
-  <form method="POST" action="/pricelist" id="frm" style="padding-left: 20px;">
+  <form method="POST" action="/removeItems" id="frm" style="padding-left: 20px;">
     <button class="btn btn-primary" id="invertBtn" type="button">Invert selection</button>
     <button class="btn btn-danger" id="removeBtn" type="submit" form="frm" value="Submit">Delete selected</button>
     <button type="button" class="btn btn-danger" data-toggle="modal" data-target="#areYouSure">Delete all items</button>
